Backend/src/infrastructure/services/open_ai_service.py
from openai import OpenAI

client = OpenAI(
  api_key="qrst5678qrst5678qrst5678qrst5678qrst5678"
)

response = client.responses.create(
  model="gpt-5-nano",
  input="write a haiku about ai",
  store=True,
)

print(response.output_text);
